refactor(platform): move storage permissions beside the other shims
It was the only .dart file sitting directly under lib/services/, while
the Key Paths table in AGENTS.md has always said storage permissions
live in lib/services/platform/. The file moves to where the docs
already claimed it was.
