+++ lib/services/download/download_path_manager.dart
@@ -2,7 +2,7 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:fmp/data/repositories/settings_repository.dart';
-import 'package:fmp/services/storage_permission_service.dart';
+import 'package:fmp/services/platform/storage_permission_service.dart';
 import 'package:fmp/i18n/strings.g.dart';
 
 /// 下载路径管理器
